requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-WTF==1.2.2
Flask-Migrate==4.1.0
Flask-Mail==0.10.0
Pillow==11.3.0
email-validator==2.2.0
WTForms==3.2.1
bcrypt==4.3.0
itsdangerous==2.2.0
